Create single-use trait to collect HydeKernel Filesystem calls
