+++ packages/framework/src/HydeKernel.php
@@ -34,6 +34,7 @@ class HydeKernel implements HydeKernelContract, Arrayable, \JsonSerializable
 
     use Foundation\Concerns\ImplementsStringHelpers;
     use Foundation\Concerns\ForwardsHyperlinks;
+    use Foundation\Concerns\ForwardsFilesystem;
 
     protected static HydeKernel $instance;
 
@@ -134,66 +135,6 @@ class HydeKernel implements HydeKernelContract, Arrayable, \JsonSerializable
         }
 
         return $this->routes;
-    }
-
-    public function path(string $path = ''): string
-    {
-        return $this->filesystem->path($path);
-    }
-
-    public function vendorPath(string $path = ''): string
-    {
-        return $this->filesystem->vendorPath($path);
-    }
-
-    public function copy(string $from, string $to): bool
-    {
-        return $this->filesystem->copy($from, $to);
-    }
-
-    public function touch(string|array $path): bool
-    {
-        return $this->filesystem->touch($path);
-    }
-
-    public function unlink(string|array $path): bool
-    {
-        return $this->filesystem->unlink($path);
-    }
-
-    public function getModelSourcePath(string $model, string $path = ''): string
-    {
-        return $this->filesystem->getModelSourcePath($model, $path);
-    }
-
-    public function getBladePagePath(string $path = ''): string
-    {
-        return $this->filesystem->getBladePagePath($path);
-    }
-
-    public function getMarkdownPagePath(string $path = ''): string
-    {
-        return $this->filesystem->getMarkdownPagePath($path);
-    }
-
-    public function getMarkdownPostPath(string $path = ''): string
-    {
-        return $this->filesystem->getMarkdownPostPath($path);
-    }
-
-    public function getDocumentationPagePath(string $path = ''): string
-    {
-        return $this->filesystem->getDocumentationPagePath($path);
-    }
-
-    public function getSiteOutputPath(string $path = ''): string
-    {
-        return $this->filesystem->getSiteOutputPath($path);
-    }
-
-    public function pathToRelative(string $path): string
-    {
-        return $this->filesystem->pathToRelative($path);
     }
 
     /**
